Move ExtendedGraphEdge to another file.
